adding refactor to environment vars to come for CONFIG
