+++ app/components/error.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react"
 import type { MetaFunction } from "remix";
 import { Banner, LanguageToggle } from "./";
-import { IMainPageProps } from "interfaces/main.interface";
+import { IMainPageProps } from "../interfaces/main.interface";
 
 export const meta: MetaFunction = () => ({
   charset: "utf-8",
