+++ app/entry.client.tsx
@@ -2,7 +2,7 @@ import i18next from "i18next";
 import { hydrate } from "react-dom";
 import { I18nextProvider, initReactI18next } from "react-i18next";
 import { RemixBrowser } from "remix";
-import { defaultLocale, supportedLanguages } from '../data/constants';
+import { defaultLocale, supportedLanguages } from './data/constants';
 
 i18next
   .use(initReactI18next)
