+++ app/i18n.server.ts
@@ -1,6 +1,6 @@
 import { RemixI18Next, FileSystemBackend } from "remix-i18next";
 import { createCookie } from '@remix-run/server-runtime'
-import { defaultLocale, supportedLanguages } from '../data/constants';
+import { defaultLocale, supportedLanguages } from './data/constants';
 
 const backend = new FileSystemBackend("./public/locales");
