+++ app/interfaces/action.interface.ts
