+++ app/interfaces/processing-statement.interface.ts
@@ -1,0 +1,5 @@
+export type IProcessingStatement = {
+  documentNumber: string;
+  userReference: string;
+  startedAt: string;
+};
