+++ app/interfaces/storage-document.interface.ts
@@ -1,0 +1,5 @@
+export type IStorageDocument = {
+  documentNumber: string;
+  userReference: string;
+  startedAt: string;
+};
