+++ app/routes/create-catch-certificate/catch-certificates.tsx
@@ -1,10 +1,10 @@
 import { ReactElement } from "react";
 import { json, useLoaderData } from "remix";
 import { Button, BUTTON_TYPE } from "@capgeminiuk/dcx-react-library";
-import { IDashboardData } from "../../../interfaces/catch-certificate.interface";
+import { IDashboardData } from "../../interfaces/catch-certificate.interface";
 import { ProgressTable, CompleteTable, NotificationBanner } from "~/components";
-import { Journeys } from "../../../data/constants";
-import CONFIG from "../../../config/config";
+import { Journeys } from "../../data/constants";
+import CONFIG from "../../config";
 
 export const loader = async () => {
   const res = await fetch(
