+++ app/routes/processing-statements/processingStatement.ts
@@ -1,10 +1,6 @@
-export type ProcessingStatement = {
-  documentNumber: string;
-  userReference: string;
-  startedAt: string;
-};
+import { IProcessingStatement } from "~/interfaces/processing-statement.interface";
 
-export function getProcessingStatements(): ProcessingStatement[] {
+export function getProcessingStatements(): IProcessingStatement[] {
   return [
     {
       documentNumber: "GBR-2022-PS-E8F37B0E8",
