+++ app/routes/storage-documents/index.tsx
@@ -1,15 +1,16 @@
 import { ReactElement } from "react";
 import { json, useLoaderData } from "remix";
-import { StorageDocument, getStorageDocuments } from "./storageDocument";
+import { getStorageDocuments } from "./storageDocument";
 import { ProgressTable } from "~/components";
-import { Journeys } from "../../../data/constants";
+import { IStorageDocument } from "~/interfaces/storage-document.interface";
+import { Journeys } from "../../data/constants";
 
 export const loader = async () => {
   return json(getStorageDocuments());
 };
 
 const Dashboard = (): ReactElement => {
-  const certificates = useLoaderData<StorageDocument[]>();
+  const certificates = useLoaderData<IStorageDocument[]>();
   return (
     <ProgressTable 
       certificates={certificates}
