--- interfaces/notification.interface.ts
@@ -1,4 +1,0 @@
-export interface INotification {
-  title: string;
-  message: string;
-};
